Refactor: initial module separation
